Removed the temp names
