+++ GameLib/Coral.py
@@ -1,5 +1,5 @@
 import pygame as pg
-from GameLib.Sprite.TempMultiAnimatedSprite import MultiAnimatedSprite
+from GameLib.Sprite.MultiAnimatedSprite import MultiAnimatedSprite
 
 class Coral(MultiAnimatedSprite):
     def __init__(self, game, health, coords, damage, path, animation_time, *args):
