+++ GameLib/MapSlot.py
